Remove color customizations
Custom colors only apply to the adl theme. These customizations
cause numerous issues in functionality of the color picker page,
and also caused issues for new tenants that were not using the
adl theme.

A new method to set custom colors needs to be implemented where
it only is applied to the adl theme.