Use PDF.js
Removes some of custom logic for when to use universal viewer.
Even if a work has been split to child works it now uses the pdf.js viewer
if that is the setting.

Some overrides still exist, to allow the viewer to display for a url or an
attached file.