Up job priority for Hyrax::QueuedIndexingJob
By default, the Hyrax::QueuedIndexingJob is set to
a priority of 0. Along with the number of jobs handled
before rescheduling and the delay between jobs, this can
result in the queue getting too backed up.

Upping the priority higher than the migration jobs lets
the indexing jobs jump ahead and prevents them from
getting stuck behind the migration jobs which were scheduled earlier.