🎁 Exclude collection from catalog search results
This commit updates the submodule that will add the ability to exclude a
collection from the catalog search results.

A checkbox to `Hide from catalog search` is added to the Discovery tab
of the collection edit form. Checking the box will set the
`hide_from_catalog_search` attribute on the CollectionResource to true
which will filter it from the search results.

The `hide_from_catalog_search` attribute is added to the collection
resource metadata.

Ref:
- https://github.com/notch8/adventist_knapsack/issues/892