🐛 Use post method for nested collection queries
This commit updates the Hyku submodule to pull in the latest version of
Hyrax 5.0-flexible that includes a bug fix in the Nested Collection
Controller. Since Nested Collection Controller is not based on catalog
controller it does not inherit the POST behavior we added there. This
causes the `parent_and_child_can_nest?` to do a GET request, which
fails when too many ids are included in the request. In this fix, we add
an explicit setting of `http_method` prior to doing this query.

Authored by @laritakr

Ref:
- notch8/adventist_knapsack#1004